pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "weekly-lists"
version = "0.1.0"
description = "Generate weekly assignment lists from insurance audits"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "Proprietary"}
dependencies = [
    "pandas>=1.3,<3.0",
    "pyyaml>=5.4",
    "openpyxl>=3.0",
    "matplotlib>=3.5",
]

[project.scripts]
weekly-lists = "run_weekly:main"

[tool.setuptools]
packages = ["src"]
package-dir = {"src" = "src"}
py-modules = ["run_weekly"]
